fz.command_on_presence
